feat: converting text conversion functions lib to typescript
