+++ src/lib/conversions/index.ts
@@ -1,0 +1,10 @@
+export { default as strOct } from './strOct';
+export { default as stringBin } from './stringBin';
+export { default as strHex } from './strHex';
+export { default as strDec } from './strDec';
+export { default as reverseStr } from './reverseStr';
+export { default as escapeString } from './escapeString';
+export { default as uri } from './uri';
+export { default as uriComp } from './uriComp';
+export { default as stringLower } from './stringLower';
+export { default as stringUpper } from './stringUpper';
