+++ src/lib/conversions/stringLower.ts
@@ -1,0 +1,5 @@
+export default class StrLower {
+  encode(input: string) {
+    return input.toLowerCase();
+  }
+}
