+++ src/lib/conversions/stringUpper.ts
@@ -1,0 +1,5 @@
+export default class StrUpper {
+  encode(input: string) {
+    return input.toUpperCase();
+  }
+}
