+++ src/lib/conversions/uri.ts
@@ -1,9 +1,9 @@
 export default class Uri {
-  encode(input) {
+  encode(input: string) {
     return encodeURI(input);
   }
 
-  decode(input) {
+  decode(input: string) {
     return decodeURI(input);
   }
 }
