--- src/utils/escapeString.js
@@ -1,9 +1,0 @@
-export default class EscapeStr {
-  encode(input) {
-    return escape(input);
-  }
-
-  decode(input) {
-    return unescape(input);
-  }
-}
