--- src/utils/stringLower.js
@@ -1,5 +1,0 @@
-export default class StrLower {
-  encode(input) {
-    return input.toLowerCase(input);
-  }
-}
