--- src/utils/stringUpper.js
@@ -1,5 +1,0 @@
-export default class StrUpper {
-  encode(input) {
-    return input.toUpperCase(input);
-  }
-}
